+++ packages/amplify-ui-components/src/components/amplify-select/amplify-select-interface.ts
@@ -1,9 +1,10 @@
 interface SelectOption<T> {
-  label: string;
-  value: T;
+	label: string;
+	value: T;
 }
 
-interface SelectOptions<T extends string | number> extends Array<SelectOption<T>> {}
+interface SelectOptions<T extends string | number>
+	extends Array<SelectOption<T>> {}
 
 export type SelectOptionsString = SelectOptions<string>;
 export type SelectOptionsNumber = SelectOptions<number>;