+++ packages/aws-amplify-react-native/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.1](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-react-native@5.0.0...aws-amplify-react-native@5.0.1) (2021-05-26)
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+
+
+
+
+## [4.3.3](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-react-native@4.3.2...aws-amplify-react-native@4.3.3) (2021-05-06)
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+
+
+
+
 ## [4.3.2](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-react-native@4.3.1...aws-amplify-react-native@4.3.2) (2021-03-12)